refactor(ai/test): split domain_planner_orchestrator_test by concern
Replace the monolithic domain_planner_orchestrator_test.dart with five
focused files: core default-API paths, fake army-move injection,
economy/naval/research via fake API, diplomacy branches, and build
selection (cargo preference + determinism).

Refs #2289
